# Keep databases open