@citronalco 2022-09-21
Regarding [#issuecomment-1253285415:](https://github.com/CastagnaIT/plugin.video.netflix/issues/1401)
Using a global might be considered bad practice, but I don't know how to handle this properly in a better way:

Using the "global", the system's libsqlite3 threadsafety is exactly checked a single time.
If I put the check into `init` of SQLiteDatabase (db_base_sqlite.py) and `cache_management.py`, the check gets executed every time a new database object gets created. I counted 5 times. This might not hurt performance that much, but it's superflous.
Is there a place where I can store this system capability, globally?