@CastagnaIT 2022-09-26
i have done a some tests on Windows and seem ok
i will do test on android next days, for now i leave some comments review for code style changes